déplace l'import de la liste vers un fichier spécifique
